requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# LoRA implementation
peft>=0.4.0

# Laplace approximation
laplace-torch>=0.1.0
backpack-for-pytorch>=1.5.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Metrics and evaluation
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
datasets>=2.14.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
